[feat] Create new player pop up stage
When the create player button in ChoosePlayer is clicked, this will
bring up the create player stage and the user will have to close the pop
up before it can interact with the app stage again. Created a helper
function to configure the modal for the pop up stages to reduce
duplicated code between files.