[feat] Create and delete txt files when profile is created and deleted
When a new player is created, a UserProfile instance is created which
contains all stats fields. These fields are stored in a txt file while
can be found in "/resources/data". Each UserProfile instance is also
mapped in a hashmap via SceneManager so that the profile instance can be
accessed in all app pages. Deleting will delete the txt file and the
user instance from the hashmap.