[feat] UserProfile helper function, reformatting and switch users
1. Added a writeData() function so that when fields are updated, this
can be called and it will write the new data to the existing user file.
2. ChoosePlayer is now the startup root to force users to choose/create
a user before starting any game.
3. Users can now switch to other users via the menu -> choose player.
Not visually seen but there is a field in the SceneManager "mainUser"
which can be used to get the UserProfile instance of the selected user
in the profile map.